New struct NetworkV6, rename struct Network -> NetworkV4.
This is due to there being a different in the type of the Netmask value,
when it is returned from the API server. For IPv4 networks, the netmask
is returned as a string. However, the netmask is returned as an int for
IPv6 networks.

Add tests for the new NetworkV6 type.