Added support for delete by tag and showing the confirmation list on delete by tag
1. Delete by tag was not supported for the command, although there was
code to do so. Added the —tag-name parameter to the delete command so
that delete by tag can actually be done.
2. I also made it so that instead of calling DeleteByTag directly,
first a list is shown, and then we delete those droplets by id if the
user confirms. The code for looping through to do the delete is in the
common deleteFn used by both delete by name/id and delete by tag. This
way the functionality is the same regardless of if doing a delete by
tag or a delete by id.