vendor: Set versions for viper and cast, remove go-spew
go-spew is a transitive dependency of testify, but we don't depend on it
directly so shouldn't need to list it in Gopkg.toml.

Set a version for viper and require a specific version of
github.com/spf13/cast that works with the viper version. I'm not really
clear why this is required, but running `dep ensure` without setting
this results in a broken build because cast doesn't get updated and the
existing version in the vendor directory lacks functions used by viper.