kubernetes: Avoid upgrading to a new minor version automatically
When doing an upgrade to the "latest" release, we want to still stay
within the same minor version. This will reduce the chance of
accidentally doing a minor version upgrade, which is nice since minor
upgrades are somewhat riskier than patch version upgrades.

If the "latest" version is requested, check the cluster's current minor
version and only consider releases of the same minor version when
selecting the version to use for upgrade.