Ensure kubeconfig exec is context-aware
Background: When an operation via doctl is invoked that writes to the
user's kubectl config (for example,
`doctl kubernetes cluster kubeconfig`), the `--context` flag is used by
doctl, but **not** kubectl.

This problematic, because we use a command called `exec-credential` that
grabs certificates on the fly from DO Kubernetes. This `exec-credential`
call is invoked by kubectl, so if kubectl is not aware of the value of
the `--context` flag passed by the user during the kubeconfig command,
then `exec-credential` will not use that context either, causing the
kubectl command to fail unexpectedly.

This change updates the `mergeKubeconfig` function responsible for
updates to the kubectl config to always pass a `--context` flag. If
there is no context set, it uses the default context, otherwise it uses
the one set by the command.