Read auth context from config if not set
Previously, if the global `Context` var was not set, we'd fall back to
the default context when writing to the user's kubeconfig. This would be
acceptable behaviour _if_ `Context` was always set to either the value
of the context flag, or the context in the user's config file, however,
`Context` only stores the value of the flag, and ignores the value of
the user's config.

This changes the way we determine the current auth context when writing
to the user's config file by first checking the value of the flag, and
then falling back to the value set in the user's config.