Write JSON for nil slices as [] instead of null
Fixes #479 (and a few other similar cases when I tested manually).

According to Go's JSON marshalling logic, a nil slice is rendered as `null` in
JSON. Due to some of the filtering we do on the returned API results, there
are cases where we send a pointer to a struct containing only a nil slice to
the displaying logic. This ends up being rendered as `null`. To correct this,
I've added a check to the JSON displaying logic for any `Displayable` type
that writes `[]` instead of `null` for this particular case.

In order to write a test for this, I decided to move the TestConfig from the
doctl_test package to the doctl package. This let me import it in the
commands_test package to reuse it, and should encourage further reuse.