Fix distfile basename
Before this change, the basename would always be generated to contain
the string `.exe`. This fixes the script in a shellcheck-compliant way
so that we're back to the old behaviour.

It also fixes the error introduced by the `u` flag in `set -euo
pipefail` when the script is invoked without a `label` or set
`SKIPBUILD` so that those fall back to an empty default value.