Add command to list auth contexts
Closes #556

Adds a basic command to list all auth contexts by reading from the
user's doctl config file. The current auth context is shown with the
word `(current)` next to it.

Testing this truly effectively is impossible due to our global Viper
instance. I could create some kind of function to retrieve the config
data, but that feels like creating too much of a mess just to make this
testable, and sidesteps the actual problem of global Viper.

That said, I've added a relatively exhaustive test to the displaying
logic, which I think is the most important part, so I feel like this is
adequately covered.