displayer no longer relies on a config
- we want to break the dependency that many objects
in doctl have on some form of config. This PR does
that by taking ONLY the values off of the config we
need to dictate display.
- DisplayText has also become a public method as we
would eventually like to remove the need for a displayer
all together BUT we want to attempt to do that incrementally.
A follow on PR will hopefully now use this public method from
within a command itself. Commands have a 3 jobs, parse their
specific config, run their action, and render their output
error.