nuke NSRoot from space!
Co-Authored-By: Zachary Gershman <zachgersh@gmail.com>

In addition to running clean on the tests, we tried out the code
with:
- doctl auth init
- doctl compute droplet list
- doctl k8s cluster list
- doctl k8s cluster kubeconfig save
and a few others. (The k8s commands are particularly of interest as
there are interactions between the doctl config and kubeconfig.)

We tried the above command with command line options and using the
values from the doctl config file.