Allow users to switch contexts
Removes the default value from the flag parsing call that sets the
global `Context` variable, because for some reason that's the value we
use in the commands *if it's set* to decide which context to use to pull
the authentication token from the user's config.

Also adds an integration test to check for this case in order to prevent
future regressions.
Co-authored-by: Zach Gershman <zachgersh@gmail.com>