adds --trace flag test to projects delete
Why is it on this command? Easier to test as there isn't
THAT much that can come back from the command.

The parsing / matching here looks crazy BUT after trying
a couple of different things I felt like the split and
three separate loops was the easiest to parse at a glance
instead of trying to dedup the test.