Accept a port of zero for SRV records
Updates the create and edit domain records commands to send a value of
zero when the port is set to zero via a flag, and to not send it
otherwise.

This required both of these commands not to use `godo`, in order to save
us from having to make a breaking change in `godo` to update the port
field to be a int pointer instead of just an int.

In the future, this commit should be reverted once we're able to release
a v2 of `godo` that addresses this issue properly.