add `doctl auth run` subcommand
`doctl auth run COMMAND [args]` will run a subprocess with the current
access token placed into the DIGITALOCEAN_TOKEN environment variable
of the subprocess environment.

This is useful to invoke terraform or other automation scripting
that needs to make DigitalOcean API calls without having to
reinvent how to securely pass the API token to the subprocess.