consolidate logic, droplet create --tagName(s)
Per @andrewsomething

"A bit of context: For Droplet --tag-name first checks that the tag
exists since, when first introduced, that was a prerequisite. If the
user had not already created that tag, the Droplet create request
would fail. These days, we will automatically create the tag for the
user if they attempt to create a Droplet using a tag that did not
previously exist. So that check is now redundant and just adds an
unnecessary API call."

In fact, it adds three unnecessary API calls!

We should look at pulling --tagName from the docs at a later time.