Add the doctl mascot
Adds the `doctl` mascot. I've updated the README to display the
`doctl` project name, the badges, and the mascot so that they're
centered. I copied the formatting used by React Native's README,
so that we're using a formatting approach that is reliable.

I've saved the mascot in our team's shared Space with the CDN
enabled, so that we don't need to unnecessarily increase `doctl`'s
total file size with an image.