Upgrade doctl to Go 1.14
Upgrades doctl, as well as our build infrastructure, to Go 1.14.

You might notice that the `vendor/modules.txt` file looks strange. In Go
1.14, the release notes state the following:

> When -mod=vendor is set (explicitly or by default), the go command now
> verifies that the main module's vendor/modules.txt file is consistent
> with its go.mod file.

Our `vendor/modules.txt` file was out of date, so I had to update it
with `go mod vendor`. As a result of that, the `go.mod` file was
changed, hence the new (and weird looking) comments in there.

I've also verified that the new Docker image and Snap base exist.