registry: Add a command to get a Docker config (#798)
* registry: Add a command to get a Docker config

Add a command, `doctl registry docker-config`, that prints a Docker config. This
is useful for users who want to configure a third-party tool such as a build
system to access their registry, and mirrors the functionality offered in the
control panel.

* integration: Add a test for `doctl registry docker-config`

* registry: Make the registry docker-config integration test safer

Avoid using a global by writing a different response for read-only
vs. read-write requests and asserting appropriately.

Co-authored-by: Andrew Starr-Bochicchio <andrewsomething@users.noreply.github.com>