go.mod: remove uneeded replace (#983)
`go mod why -m github.com/strechr/objx` reveals that this package doesn't
actually get built or included in the doctl command. This replace was
added to force the objx package to v0.2.0 because of security vuln. `go
list -mod=mod -m github.com/stretchr/objx` reveals that v0.2.0 is the
version that would be used anyway if it was included.

Removing this also enables building of doctl via the new go intall
pkg@version command which rejects replace statements.