Fix auth key mangling when switching contexts (#998)
"Fixes" a bizarre issue where running `doctl auth switch`
mangles all auth contexts that contain a `.` character.

Prior to this change, when the command was run, it appears
that Viper decides that anything containing a `.` indicates
that the value should be split at the `.`, and then anything
after it should become the key of a nested entry in a map
in the config YAML file.

I have no idea why this fix works, but it does. I'm still
looking into the root cause, but I figure that we might as
well get this change out.