Update domain record without changing TTL (#1020)
Given an existing domain record with an non-default TTL, a user can not
update a field on the record without TTL changing to 1800 unless the
user remembers to specify the desired TTL on each update.  This patch
changes the default value of the TTL argument from 1800 to 0 for domain
record updates which preserves the current TTL of the record.