Add support for finding the sandbox namespace
This allows doctl sandbox connect to be invoked with no argument.
The algorithm is likely to change as we make a decision about how to
provision sandbox namespaces for customers in EA.  For now, most of us
internally use the alpha console to get namespaces and that continues
to work.