Ensure that sbx init --language value is supported (#1119)
Previously, a language would be accepted if it was in a static list of
languages for which we know how to generate samples.  But, not all
languages in that list necessarily have runtimes in the backend, which
was leading to a situation where the developer could initialize a local
sandbox area that could not be deployed.