Fix potential loop in `apps update --wait` (#1143)
When there are no changes to make, it's possible for there to never be
an `InProgressDeployment` returned from the API, due to the deployment
finishing between the call to `update` and the subsequent call to `Get`
the status. As a fix, we now fetch the latest deployment for the app,
rather than relying on `InProgressDeployment`.

There's no ideal solution here, since the deployments cannot be
correlated with the request that triggered them. The implemented
approach relies on deployments being observable immediately after the
triggering request, and could be confused in the event of concurrent
deployment requests (which is probably unlikely).

Co-authored-by: Andrew Starr-Bochicchio <andrewsomething@users.noreply.github.com>