Move the sandbox feature to the main branch (#1154)
* Establish sandbox management commands

* Add autogenned stuff, not edited yet.

All the new trees are being picked up and seem to work (generally)

* Fix table formatting for list commands

* Fix gofmt issues

* Ignore artifacts of the sandbox build/upload cmd.

* Minor changes, adjust to new plugin output options

* Eagerly convert result.Error to golang error

Also adjusts output processing to be more suitable when there are
multiple categories of output and simplifies its logic.

* Fixes to get 'sandbox-functions watch' to work

Note that we plan to re-run codegen once more and command names will
change in that process.   That hasn't been done yet.  But, the watch
function works.  There's an orthogonal bug in 'nim project watch' that
needs to be fixed before symlinks inside node_modules will be
tolerated.  Until then, symlinks can cause the watcher to oscillate.

* Re-organize command tree, address usage texts

Still some more work to do for activations.  The 'activations logs'
subtree is mysteriously broken somewhere in the handoff to AIO.

* Allow debugging output to flow alongside result

* Two fixes. (1) Finish fixing usage texts.

(2) Correct handling of 'package' and 'deployed' flags for
'activation logs'.

* Fix up some last details for sandbox-extra cmds

* Hide sandbox command unless installed

* Remove unneeded change from initial PR

* Fix bug in --param flag handling for fn invoke

* Mock out SandboxSubprocess.

* Expose way to override sandbox install checks for mocks.

* sandbox: Add additional tests. (#1107)

* added link for size slugs (#1104)

Co-authored-by: Cesar Garza <scotch.neat@live.com>

* Run PR workflow for feature branches. (#1108)

* sandbox: Add additional tests.

* PrintSandboxTextOutput should use CmdConfig.Out

Co-authored-by: Dan Brian <daniel.brian@gmail.com>
Co-authored-by: Cesar Garza <scotch.neat@live.com>

* WIP for versioning and update

* Repair broken unit tests

* Fix logic in upgrade.  Adjust tests.

* Fix doctl sandbox connect command

WIP for consolidating install and connect.  Step 1 is to get connect
working as before.

* Add support for finding the sandbox namespace

This allows doctl sandbox connect to be invoked with no argument.
The algorithm is likely to change as we make a decision about how to
provision sandbox namespaces for customers in EA.  For now, most of us
internally use the alpha console to get namespaces and that continues
to work.

* Fix the sandbox connect test

* Additional test (another path through sbx connect)

* Apply suggestions from code review

Co-authored-by: Andrew Starr-Bochicchio <andrewsomething@users.noreply.github.com>

* Fixes for PR #1111 to pass glint checks (new code)

* Fix lint errors in new sources on feature/sandbox

* Correct behavior for 'sandbox actv logs --poll'

* Use --follow for actv logs, consistent with doctl

This came up in the review of the PR and is being added as part of it.

* Add an 'undeploy' command.

* Fix usage text for undeploy

* Add unit tests for the new undeploy command

* undeploy: Make test relationships more explicit.

* Change node version to 16.13.0

Also change the sandbox nim version to 2.3.7.  Older releases will
almost certainly work fine on 16.x but starting with 2.3.7 we are
explicitly testing with it.

* Ensure that sbx init --language value is supported (#1119)

Previously, a language would be accepted if it was in a static list of
languages for which we know how to generate samples.  But, not all
languages in that list necessarily have runtimes in the backend, which
was leading to a situation where the developer could initialize a local
sandbox area that could not be deployed.

* Adopt nim 2.3.9, which fixes two bugs (#1121)

* sandbox: Use Go native libs to extract archives. (#1122)

* sandbox: Support Windows node downloads.

* Use exe file extension on Windows.

* Improvements to flag handling and include/exclude for 'web' (#1124)

* Incorporate newer nim version with bug fix (#1126)

* Use new simplified API to get sandbox namespace

This removes the need for the namespace to be added via the cloud
console or administratively prior to connecting.  The connect operation
should "always" work (in beta, only if you have the feature-flip, in GA
it should work for all customers).

* Fix unit tests and mocks for API change

* Maintain sandbox credentials on per-DO-token basis

* Fix blank line anomaly in output

* Use hashing instead of truncation for tokens

* Additional testing of credential handling. (#1136)

* Fix bug in activations logs --follow (#1137)

* Improvements to error output display

Improves the output from deploy so that there is more context for the
actual error.  Also ensures that a non-zero exit code is posted on a
deploy error,

* Increment to nim version 3.0.4

The newer 'nim' invokes newer configurations of the builder system
functions, allowing more efficient and secure remote builds.

* Support e2e tests with added env vars (#1148)

* Store API host with different final domain segment (#1149)

* Adds check for connected status before plugin call (#1150)

The check is bypassed for the sandbox connect command (only).

* Fix bug in naming the node executable on windows

It is always 'node.exe' not 'node'.  This was causing 'sandbox upgrade'
to fail.

* Avoid "cross file system" errors during install

* Don't pass zero args to deploy or watch (always 1)

Otherwise, nim will just dump help.  For deploy, this causes a JSON
parse error.  For watch, it causes inappropriate nimbella-branded stuff
in the watch window.

* Increase nim version and sandbox plugin version

Picks up latest bug fixes from nim.  New plugin code adds to branding
so that the output of doctl sbx watch doesn't look so "nim-ish".

* Fix dangling import error

* Make serverless and sis aliases of sandbox

* Fix two bugs: handling of arch=386, sbx init issue

Co-authored-by: Andrew Starr-Bochicchio <a.starr.b@gmail.com>
Co-authored-by: Andrew Starr-Bochicchio <andrewsomething@users.noreply.github.com>
Co-authored-by: Dan Brian <daniel.brian@gmail.com>
Co-authored-by: Cesar Garza <scotch.neat@live.com>