uptime checks: add crud methods (#1386)
Tried to see how far I could get adding support for uptime checks and
alerts in a day, and this is what I could do.

This is still missing:

  * Integration tests,
  * More in-depth tests for the `create` check operation, since I added
    a bit of extra functionality over the default API interaction,
  * The `get state` method for checks. It has a different displayable
    type, so it was too much work for today,
  * All of the uptime alerts, since it's a separate set of endpoints.

Not sure if we want to merge as-is, but I figured it's probably better
than not having the functionality. I'll try to complete it on the next
free Friday.