kubernetes: remove backquotes from flag descriptions
Remove backquotes from flag descriptions in the kubernetes subcommands,
because those backquotes cause pflag's usage parser to extract them and
treat them as a flag name to try to determine the data type of the flag.
The practical effect of this was that the help was borderline unreadable
because the description was forced so far right and narrow.