APPS: Add an --update-sources parameter to create and update app (#1622)
This adds a parameter named `--update-sources` to the `apps create` and `apps update` commands. If set, the respective update of the app spec also causes the underlying sources to be reevaluated and their respective latest state to be reflected in the created deployment.

On the `apps create` flow, this only applies to `--upsert` cases.